backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
marshmallow==3.20.1
pydantic==2.5.2
gunicorn==21.2.0
pytest==7.4.3 